+++ custom_components/alexa_assist_bridge/const.py
@@ -5,9 +5,11 @@ DOMAIN = "alexa_assist_bridge"
 CONF_ALEXA_SKILL_ID = "alexa_skill_id"
 CONF_ALLOW_DEBUG_REQUESTS = "allow_debug_requests"
 CONF_AGENT_ID = "agent_id"
+CONF_ASSISTANT_NAME = "assistant_name"
 CONF_ENDPOINT_ID = "endpoint_id"
 CONF_LANGUAGE = "language"
 
 DEFAULT_AGENT_ID = "conversation.home_assistant"
+DEFAULT_ASSISTANT_NAME = "Nabu"
 DEFAULT_LANGUAGE = "en"
 DEFAULT_ALLOW_DEBUG_REQUESTS = True